app_python_backend/requirements.txt
sqlalchemy
pymysql
prometheus_client
Flask
flask_restful
python-dateutil
#pandas